dict 클래스 확장 및 base class로 적용
